Optimization & BuildingInventory Update
-Оптимизация скриптов строений и юнитов
-Добавлен префаб фермы(построить ферму на кнопку 5)
-Переработан инвентарь строений, теперь все здания имеют хранилище
-Теперь есть общий класс с массой всех возможных ресурсов, доступный всему проекту
-Добавлена возможность дистанционного переноса ресурсов в здание(выберите рабочего и нажмите кнопку G на здание)(ResourcesTransferWindowController должен быть вынесен на сцену)(Настройка расстояния есть, но она работает некорректно из-за неправильной работы метода Distance в скрипте hexgrid, так что пока просто поставьте значение побольше)

+++ Assets/Scripts/Buildings/FirstFactionFacilities.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 
 public class FirstFactionFacilities : FacilityDescription
 {
-    public float FoodConsumption;
-    public float OreConsumption;
-    public float WoodConsumption;
-    public float LightConsumption;
+    public int FoodConsumption;
+    public int OreConsumption;
+    public int WoodConsumption;
+    public int LightConsumption;
 }
